requirements.txt
bottle
python-memcached
paste
